Fix routes and processing creation from template
